feat: Add Makefile
With init, compile, upload, monitor and clean targets plus simple detection of the port
for upload and monitor. Does not handle errors well wi a board isn't attached and
libraries are placed in the global defaults. Also, make compile dependent state of the
SKETCH and the build target.

More work in the future but this is basic functionality.