feat: Add `help` target
Also, the target and alias sections are generated by the `awk` in help
so it help is DRY and automatic. ChatGPT had a great idea there and it
also caught the unquoted variable expansion!

I've chosen to have no blank line between target and alias so the
association is more obvious, but I could change my mind in the future :)